# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "post": {
        "description": "Permite crear un nuevo producto/repuesto",
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductsCreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsCreateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "409": {
            "description": "El recurso ya existe"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Crear nuevo producto",
        "tags": [
          "Products"
        ]
      },
      "get": {
        "description": "Permite obtener una lista paginada de productos",
        "operationId": "ProductsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Productos encontrados correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Listar productos",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "description": "Permite obtener un producto específico por su ID",
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsCreateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Obtener producto por ID",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "description": "Permite actualizar un producto existente",
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductsUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsUpdateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "409": {
            "description": "El recurso ya existe"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Actualizar producto",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "description": "Permite eliminar un producto (soft delete)",
        "operationId": "ProductsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Eliminar producto",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/api/log-emails/{id}": {
      "get": {
        "description": "Permite obtener un registro de envío de email específico",
        "operationId": "LogEmailsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log de email encontrado correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Recurso no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Obtener log de email por ID",
        "tags": [
          "LogEmails"
        ]
      }
    },
    "/api/log-emails": {
      "get": {
        "description": "Permite obtener una lista paginada de logs de envío de emails",
        "operationId": "LogEmailsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logs de emails encontrados correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "summary": "Listar logs de emails",
        "tags": [
          "LogEmails"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sistema de Almacén API",
    "description": "API para gestión de almacén - Productos, Cotizaciones, Ventas y Proveedores",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "ProductsCreateRequestDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Filtro de aceite motor"
          },
          "descripcion": {
            "type": "string",
            "example": "Filtro de aceite para motor 1.6L"
          },
          "precio": {
            "type": "number",
            "example": 25.5
          },
          "stock": {
            "type": "number",
            "example": 50
          },
          "categoria": {
            "type": "string",
            "example": "Filtros"
          },
          "proveedorId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "nombre",
          "precio",
          "stock",
          "proveedorId"
        ]
      },
      "ProductsCreateResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Product created successfully"
          },
          "id": {
            "type": "number",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "example": "Filtro de aceite motor"
          },
          "descripcion": {
            "type": "object",
            "example": "Filtro de aceite para motor 1.6L"
          },
          "precio": {
            "type": "number",
            "example": 25.5
          },
          "stock": {
            "type": "number",
            "example": 50
          },
          "categoria": {
            "type": "object",
            "example": "Filtros"
          },
          "activo": {
            "type": "boolean",
            "example": true
          },
          "fechaRegistro": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "proveedorId": {
            "type": "number",
            "example": 1
          },
          "status_code": {
            "type": "number",
            "example": 201
          }
        },
        "required": [
          "message",
          "id",
          "nombre",
          "precio",
          "stock",
          "activo",
          "fechaRegistro",
          "proveedorId",
          "status_code"
        ]
      },
      "ProductsUpdateRequestDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Filtro de aceite motor actualizado"
          },
          "descripcion": {
            "type": "string",
            "example": "Filtro de aceite para motor 1.6L actualizado"
          },
          "precio": {
            "type": "number",
            "example": 30
          },
          "stock": {
            "type": "number",
            "example": 75
          },
          "categoria": {
            "type": "string",
            "example": "Filtros Premium"
          },
          "proveedorId": {
            "type": "number",
            "example": 1
          }
        }
      },
      "ProductsUpdateResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Product updated successfully"
          },
          "status_code": {
            "type": "number",
            "example": 200
          }
        },
        "required": [
          "message",
          "status_code"
        ]
      },
...
    }
  }
}